+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - [#6379](https://github.com/influxdata/influxdb/issues/6379): Validate the first argument to percentile() is a variable.
 - [#6294](https://github.com/influxdata/influxdb/issues/6294): Fix panic running influx_inspect info.
 - [#6382](https://github.com/influxdata/influxdb/pull/6382): Removed dead code from the old query engine.
+- [#6383](https://github.com/influxdata/influxdb/pull/6383): Recover from a panic during query execution.
 
 ## v0.12.0 [2016-04-05]
 ### Release Notes